requirements.txt
discord.py
psutil
python-dotenv
aiomcrcon
aiofiles
requests
pytz
